refactor: separate exception from validation package
