+++ src/main/java/racingcar/exception/RacingCarException.java
@@ -1,0 +1,11 @@
+package racingcar.exception;
+
+public class RacingCarException extends IllegalArgumentException {
+    private RacingCarException(ErrorMessage errorMessage) {
+        super(errorMessage.getMessage());
+    }
+
+    public static RacingCarException of(ErrorMessage errorMessage) {
+        return new RacingCarException(errorMessage);
+    }
+}
