+++ src/main/java/racingcar/validation/registry/MinimumLengthValidator.java
@@ -1,14 +1,15 @@
 package racingcar.validation.registry;
 
 import java.util.List;
-import racingcar.validation.ErrorMessage;
+import racingcar.exception.ErrorMessage;
+import racingcar.exception.RacingCarException;
 
 public class MinimumLengthValidator implements RegistryValidator {
 
     @Override
     public void validate(List<String> inputValue) {
         if (CountValueUnderOne(inputValue) != 0) {
-            throw new IllegalArgumentException(ErrorMessage.MINIMUM_NAME_LENGTH.getMessage());
+            throw RacingCarException.of(ErrorMessage.MINIMUM_NAME_LENGTH);
         }
     }
